Added match's profile to chatroom
When you are in a chatroom with another user you can now click on an information icon and see their profile! There is a slight bug, the first time you click the info icon after opening the chatroom, it does not appear and you must click it a second time bc its getting their data from firebase on the initial click. But double tapping, even the first time, works opens it instantly, so... it's a... very low priority bug imo?